sec: refactor and more testing
