+++ std/security/certificate_test.go
@@ -3,7 +3,6 @@ package security_test
 import (
 	"encoding/base64"
 	"testing"
-	"time"
 
 	enc "github.com/named-data/ndnd/std/encoding"
 	"github.com/named-data/ndnd/std/ndn"
@@ -13,36 +12,6 @@ import (
 	"github.com/named-data/ndnd/std/utils"
 	"github.com/stretchr/testify/require"
 )
-
-// Name: /ndn/alice/KEY/cK%1D%A4%E1%5B%91%CF
-// SigType: Ed25519
-const KEY_ALICE = `
-BsoHGwgDbmRuCAVhbGljZQgDS0VZCAhjSx2k4VuRzxQDGAEJFUC64F62YK0/v5z4
-fjONZO7Y4PNqy7FiDnar33uVO71FLK6Vp8GrPCkEhuODl6GBv2nUuovtO9KtHW11
-8apSS093FiIbAQUcHQcbCANuZG4IBWFsaWNlCANLRVkICGNLHaThW5HPF0Cw3Oh7
-I2jmBBxop1bIPXq292TfltVwhdbB3/yUXkKcg3BYbY6vcAhNNqrG2B+G/iHvKGsy
-DpvDtnlEN72hIeIP
-`
-
-// Name: /ndn/KEY/%27%C4%B2%2A%9F%7B%81%27/ndn/v=1651246789556
-// SigType: ECDSA-SHA256
-// Validity: 2022-04-29 15:39:50 +0000 UTC - 2026-12-31 23:59:59 +0000 UTC
-const CERT_ROOT = `
-Bv0BSwcjCANuZG4IA0tFWQgIJ8SyKp97gScIA25kbjYIAAABgHX6c7QUCRgBAhkE
-ADbugBVbMFkwEwYHKoZIzj0CAQYIKoZIzj0DAQcDQgAEPuDnW4oq0mULLT8PDXh0
-zuBg+0SJ1yPC85jylUU+hgxX9fDNyjlykLrvb1D6IQRJWJHMKWe6TJKPUhGgOT65
-8hZyGwEDHBYHFAgDbmRuCANLRVkICCfEsiqfe4En/QD9Jv0A/g8yMDIyMDQyOVQx
-NTM5NTD9AP8PMjAyNjEyMzFUMjM1OTU5/QECKf0CACX9AgEIZnVsbG5hbWX9AgIV
-TkROIFRlc3RiZWQgUm9vdCAyMjA0F0gwRgIhAPYUOjNakdfDGh5j9dcCGOz+Ie1M
-qoAEsjM9PEUEWbnqAiEApu0rg9GAK1LNExjLYAF6qVgpWQgU+atPn63Gtuubqyg=
-`
-
-var ISSUER = enc.NewStringComponent(enc.TypeGenericNameComponent, "myissuer")
-var KEY_ALICE_NAME, _ = enc.NameFromStr("/ndn/alice/KEY/cK%1D%A4%E1%5B%91%CF")
-var KEY_ROOT_NAME, _ = enc.NameFromStr("/ndn/KEY/%27%C4%B2%2A%9F%7B%81%27")
-
-var T1 = time.Date(2022, 4, 29, 15, 39, 50, 0, time.UTC)
-var T2 = time.Date(2089, 12, 31, 23, 59, 59, 0, time.UTC)
 
 func TestSignCertInvalid(t *testing.T) {
 	utils.SetTestingT(t)
